Refactor tower observer client
