+++ plugins/nf-tower/src/main/io/seqera/tower/plugin/TowerClient.groovy
@@ -52,22 +52,15 @@ import nextflow.util.ProcessHelper
 import nextflow.util.TestOnly
 import nextflow.util.Threads
 /**
- * Send out messages via HTTP to a configured URL on different workflow
- * execution events.
+ * Perform HTTP call to Seqera platform.
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
 @Slf4j
 @CompileStatic
-class TowerClient implements TraceObserverV2 {
+class TowerClient {
 
     static final public String DEF_ENDPOINT_URL = 'https://api.cloud.seqera.io'
-
-    static private final int TASKS_PER_REQUEST = 100
-
-    static private final Duration REQUEST_INTERVAL = Duration.of('1 sec')
-
-    static private final Duration ALIVE_INTERVAL = Duration.of('1 min')
 
     static private final String TOKEN_PREFIX = '@token:'
 
@@ -79,82 +72,29 @@ class TowerClient implements TraceObserverV2 {
         boolean isError() { code < 200 || code >= 300 }
     }
 
-    @ToString(includeNames = true)
-    static class ProcessEvent {
-        TraceRecord trace
-        boolean completed
-    }
-
-    private Session session
-
-    /**
-     * Workflow identifier, will be taken from the Session() object later
-     */
-    private String runName
-
-    /**
-     * Store the sessions unique ID for downstream reference purposes
-     */
-    private String runId
-
     private HxClient httpClient
 
     private JsonGenerator generator
 
-    private String workflowId
-
-    private String watchUrl
-
     private String endpoint
-
-    private ResourcesAggregator aggregator
 
     protected Map<String,String> env = System.getenv()
 
-    private LinkedBlockingQueue<ProcessEvent> events = new LinkedBlockingQueue()
-
-    private Thread sender
-
-    private Duration requestInterval = REQUEST_INTERVAL
-
-    private Duration aliveInterval = ALIVE_INTERVAL
-
-    private LinkedHashSet<String> processNames = new LinkedHashSet<>(20)
-
     private Map<String,Integer> schema = Collections.emptyMap()
-
-    private int maxRetries = 5
-
-    private int backOffDelay
-
-    private int backOffBase
-
-    private boolean towerLaunch
 
     private String accessToken
 
-    private String workspaceId
-
-    private TowerReports reports
-
     private TowerRetryPolicy retryPolicy
 
-    private Map<String,Boolean> allContainers = new ConcurrentHashMap<>()
-
-    TowerClient(Session session, TowerConfig config) {
-        this.session = session
+    TowerClient(TowerConfig config, Map env) {
         this.endpoint = checkUrl(config.endpoint)
         this.accessToken = config.accessToken
-        this.workspaceId = config.workspaceId
         this.retryPolicy = config.retryPolicy
         this.schema = loadSchema()
         this.generator = TowerJsonGenerator.create(schema)
-        this.reports = new TowerReports(session)
-    }
-
-    TowerClient withEnvironment(Map env) {
-        this.env = env
-        return this
+        if (env)
+            this.env = env
+        initHttpClient()
     }
 
     @TestOnly
@@ -162,40 +102,7 @@ class TowerClient implements TraceObserverV2 {
         this.generator = TowerJsonGenerator.create(Collections.EMPTY_MAP)
     }
 
-    @Override
-    boolean enableMetrics() { true }
-
     String getEndpoint() { endpoint }
-
-    String getWorkflowId() { workflowId }
-
-    boolean getTowerLaunch() { towerLaunch }
-
-    String getRunName() { runName }
-
-    String getRunId() { runId }
-
-    void setAliveInterval(Duration d) {
-        this.aliveInterval = d
-    }
-
-    void setRequestInterval(Duration d) {
-        this.requestInterval = d
-    }
-
-    void setMaxRetries( int value ) {
-        this.maxRetries = value
-    }
-
-    void setBackOffBase( int value ) {
-        this.backOffBase = value
-    }
-
-    void setBackOffDelay( int value ) {
-        this.backOffDelay = value
-    }
-
-    String getWorkspaceId() { workspaceId }
 
     /**
      * Check the URL and create an HttpPost() object. If a invalid i.e. protocol is used,
@@ -224,82 +131,79 @@ class TowerClient implements TraceObserverV2 {
         return "${url.protocol}://${url.authority}"
     }
 
-    protected String getUrlTraceCreate() {
+    Map traceCreate(Map req, String workspaceId){
+        return sendAndProcessRequest( getUrlTraceCreate(workspaceId), req, 'POST')
+    }
+
+    Map traceBegin(Map req, String workspaceId, String workflowId){
+        return sendAndProcessRequest( getUrlTraceBegin(workspaceId, workflowId), req, 'POST')
+    }
+
+    void traceComplete(Map req, String workspaceId, String workflowId) {
+        final url = getUrlTraceComplete(workspaceId, workflowId)
+        final resp = sendHttpMessage(url, req, 'PUT')
+        logHttpResponse(url, resp)
+    }
+
+    void traceHeartbeat(Map req, String workspaceId, String workflowId) {
+        final url = getUrlTraceHeartbeat( workspaceId, workflowId)
+        final resp = sendHttpMessage(url, req, 'PUT')
+        logHttpResponse(url, resp)
+    }
+
+    void traceProgress(Map req, String workspaceId, String workflowId) {
+        final url = getUrlTraceProgress( workspaceId, workflowId )
+        final resp = sendHttpMessage(url, req, 'PUT')
+        logHttpResponse(url, resp)
+    }
+
+    protected Map sendAndProcessRequest(String url, Map req, String method){
+        final resp = sendHttpMessage(url, req, method)
+        if( resp.error ) {
+            log.debug """\
+                Unexpected HTTP response
+                - endpoint    : $url
+                - status code : $resp.code
+                - response msg: $resp.cause
+                """.stripIndent(true)
+            throw new AbortOperationException(resp.message)
+        }
+        return parseTowerResponse(resp)
+    }
+
+    protected String getUrlTraceCreate(String workspaceId) {
         def result = this.endpoint + '/trace/create'
         if( workspaceId )
             result += "?workspaceId=$workspaceId"
         return result
     }
 
-    protected String getUrlTraceBegin() {
+    protected String getUrlTraceBegin(String workspaceId, String workflowId) {
         def result = "$endpoint/trace/$workflowId/begin"
         if( workspaceId )
             result += "?workspaceId=$workspaceId"
         return result
     }
 
-    protected String getUrlTraceComplete() {
+    protected String getUrlTraceComplete(String workspaceId, String workflowId) {
         def result = "$endpoint/trace/$workflowId/complete"
         if( workspaceId )
             result += "?workspaceId=$workspaceId"
         return result
     }
 
-    protected String getUrlTraceHeartbeat() {
+    protected String getUrlTraceHeartbeat(String workspaceId, String workflowId) {
         def result = "$endpoint/trace/$workflowId/heartbeat"
         if( workspaceId )
             result += "?workspaceId=$workspaceId"
         return result
     }
 
-    protected String getUrlTraceProgress() {
+    protected String getUrlTraceProgress(String workspaceId, String workflowId) {
         def result = "$endpoint/trace/$workflowId/progress"
         if( workspaceId )
             result += "?workspaceId=$workspaceId"
         return result
-    }
-
-    /**
-     * On workflow start, submit a message with some basic
-     * information, like Id, activity and an ISO 8601 formatted
-     * timestamp.
-     * @param session The current Nextflow session object
-     */
-    @Override
-    void onFlowCreate(Session session) {
-        log.debug "Creating Seqera Platform observer -- endpoint=$endpoint; requestInterval=$requestInterval; aliveInterval=$aliveInterval; maxRetries=$maxRetries; backOffBase=$backOffBase; backOffDelay=$backOffDelay"
-
-        this.session = session
-        this.aggregator = new ResourcesAggregator()
-        this.runName = session.getRunName()
-        this.runId = session.getUniqueId()
-
-        // send hello to verify auth
-        final req = makeCreateReq(session)
-        final resp = sendHttpMessage(urlTraceCreate, req, 'POST')
-        if( resp.error ) {
-            log.debug """\
-                Unexpected HTTP response
-                - endpoint    : $urlTraceCreate
-                - status code : $resp.code
-                - response msg: $resp.cause
-                """.stripIndent(true)
-            throw new AbortOperationException(resp.message)
-        }
-        final ret = parseTowerResponse(resp)
-        this.workflowId = ret.workflowId
-        if( !workflowId )
-            throw new AbortOperationException("Invalid Seqera Platform API response - Missing workflow Id")
-        log.debug "Platform workflow id: $workflowId; workflow url: ${ret.watchUrl}"
-        session.workflowMetadata.platform.workflowId = workflowId
-        // note: `watchUrl` in the create response requires Platform 26.01 or later
-        this.watchUrl = ret.watchUrl as String
-        session.workflowMetadata.platform.workflowUrl = watchUrl
-        if( ret.message )
-            log.warn(ret.message.toString())
-
-        // Prepare to collect report paths if tower configuration has a 'reports' section
-        reports.flowCreate(workflowId)
     }
 
     protected void initHttpClient() {
@@ -349,93 +253,10 @@ class TowerClient implements TraceObserverV2 {
         config.basicAuth(TOKEN_PREFIX + token)
     }
 
-    protected Map makeCreateReq(Session session) {
-        def result = new HashMap(5)
-        result.sessionId = session.uniqueId.toString()
-        result.runName = session.runName
-        result.projectName = session.workflowMetadata.projectName
-        result.repository = session.workflowMetadata.repository
-        result.workflowId = env.get('TOWER_WORKFLOW_ID')
-        result.instant = Instant.now().toEpochMilli()
-        this.towerLaunch = result.workflowId != null
-        return result
-    }
-
-    @Override
-    void onProcessCreate(TaskProcessor process) {
-        log.trace "Creating process ${process.name}"
-        if( !processNames.add(process.name) )
-            throw new IllegalStateException("Process name `${process.name}` already used")
-    }
-
-    @Override
-    void onFlowBegin() {
-        // configure error retry
-
-        final req = makeBeginReq(session)
-        final resp = sendHttpMessage(urlTraceBegin, req, 'PUT')
-        if( resp.error ) {
-            log.debug """\
-                Unexpected HTTP response
-                - endpoint    : $urlTraceBegin
-                - status code : $resp.code
-                - response msg: $resp.cause
-                """.stripIndent(true)
-            throw new AbortOperationException(resp.message)
-        }
-
-        final payload = parseTowerResponse(resp)
-        this.watchUrl ?= payload.watchUrl
-        session.workflowMetadata.platform.workflowUrl ?= watchUrl
-        this.sender = Threads.start('Tower-thread', this.&sendTasks0)
-        final msg = "Monitor the execution with Seqera Platform using this URL: ${watchUrl}"
-        log.info(LoggerHelper.STICKY, msg)
-    }
-
     String getAccessToken() {
         if( !accessToken )
             throw new AbortOperationException("Missing Seqera Platform access token -- Make sure there's a variable TOWER_ACCESS_TOKEN in your environment")
         return accessToken
-    }
-
-    /**
-     * Send an HTTP message when the workflow is completed.
-     */
-    @Override
-    void onFlowComplete() {
-        // publish runtime reports
-        reports.publishRuntimeReports()
-        // submit the completion record
-        if( sender ) {
-            events << new ProcessEvent(completed: true)
-            // wait the submission of pending events
-            sender.join()
-        }
-        // wait and flush reports content
-        reports.flowComplete()
-        // notify the workflow completion
-        // note: only send complete if onFlowBegin was invoked (sender is set there)
-        if( workflowId && sender ) {
-            final req = makeCompleteReq(session)
-            final resp = sendHttpMessage(urlTraceComplete, req, 'PUT')
-            logHttpResponse(urlTraceComplete, resp)
-        }
-    }
-
-    @Override
-    void onTaskPending(TaskEvent event) {
-        events << new ProcessEvent(trace: event.trace)
-    }
-
-    /**
-     * Send an HTTP message when a process has been submitted
-     *
-     * @param handler A {@link TaskHandler} object representing the task submitted
-     * @param trace A {@link TraceRecord} object holding the task metadata and runtime info
-     */
-    @Override
-    void onTaskSubmit(TaskEvent event) {
-        events << new ProcessEvent(trace: event.trace)
     }
 
     /**
@@ -444,61 +265,6 @@ class TowerClient implements TraceObserverV2 {
      * @param handler A {@link TaskHandler} object representing the task started
      * @param trace A {@link TraceRecord} object holding the task metadata and runtime info
      */
-    @Override
-    void onTaskStart(TaskEvent event) {
-        events << new ProcessEvent(trace: event.trace)
-    }
-
-    /**
-     * Send an HTTP message, when a process completed
-     *
-     * @param handler A {@link TaskHandler} object representing the task completed
-     * @param trace A {@link TraceRecord} object holding the task metadata and runtime info
-     */
-    @Override
-    void onTaskComplete(TaskEvent event) {
-        events << new ProcessEvent(trace: event.trace)
-
-        synchronized (this) {
-            aggregator.aggregate(event.trace)
-        }
-    }
-
-    @Override
-    void onTaskCached(TaskEvent event) {
-        // event was triggered by a stored task, ignore it
-        if( !event.trace )
-            return
-
-        // add the cached task event
-        events << new ProcessEvent(trace: event.trace)
-
-        // remove the record from the current records
-        synchronized (this) {
-            aggregator.aggregate(event.trace)
-        }
-    }
-
-    /**
-     * Send an HTTP message, when a workflow has failed
-     *
-     * @param handler A {@link TaskHandler} object representing the task that caused the workflow execution to fail (it may be null)
-     * @param trace A {@link TraceRecord} object holding the task metadata and runtime info (it may be null)
-     */
-    @Override
-    void onFlowError(TaskEvent event) {
-        events << new ProcessEvent(trace: event.trace)
-    }
-
-    /**
-     * Update reports file when a file is published
-     *
-     * @param destination File path at `publishDir` of the published file.
-     */
-    @Override
-    void onFilePublish(FilePublishEvent event) {
-        reports.filePublish(event.target)
-    }
 
     /**
      * Little helper method that sends a HTTP POST message as JSON with
@@ -559,169 +325,6 @@ class TowerClient implements TraceObserverV2 {
         throw new IllegalArgumentException("Unsupported HTTP verb: $verb")
     }
 
-    protected boolean isCliLogsEnabled() {
-        return env.get('TOWER_ALLOW_NEXTFLOW_LOGS') == 'true'
-    }
-
-    protected String getOperationId() {
-        if( !isCliLogsEnabled() )
-            return null
-        try {
-            if( env.get('AWS_BATCH_JOB_ID') )
-                return  "aws-batch::${env.get('AWS_BATCH_JOB_ID')}"
-            else
-                return "local-platform::${ProcessHelper.selfPid()}"
-        }
-        catch (Exception e) {
-            log.warn "Unable to retrieve native environment operation id", e
-            return null
-        }
-    }
-
-    protected String getLogFile() {
-        return isCliLogsEnabled() ? env.get('NXF_LOG_FILE') : null
-    }
-
-    protected String getOutFile() {
-        return isCliLogsEnabled() ? env.get('NXF_OUT_FILE') : null
-    }
-
-    protected Map makeBeginReq(Session session) {
-        def workflow = session.getWorkflowMetadata().toMap()
-        workflow.params = session.getParams()
-        workflow.id = getWorkflowId()
-        workflow.remove('stats')
-
-        // render as a string
-        workflow.container = mapToString(workflow.container)
-        workflow.configText = session.resolvedConfig
-        // extra metadata
-        workflow.operationId = getOperationId()
-        workflow.logFile = getLogFile()
-        workflow.outFile = getOutFile()
-
-        def result = new LinkedHashMap(5)
-        result.workflow = workflow
-        result.processNames = new ArrayList(processNames)
-        result.towerLaunch = towerLaunch
-        result.instant = Instant.now().toEpochMilli()
-        return result
-    }
-
-    protected Map makeCompleteReq(Session session) {
-        def workflow = session.getWorkflowMetadata().toMap()
-        workflow.params = session.getParams()
-        workflow.id = getWorkflowId()
-        // render as a string
-        workflow.container = mapToString(workflow.container)
-        workflow.configText = session.resolvedConfig
-        // extra metadata
-        workflow.operationId = getOperationId()
-        workflow.logFile = getLogFile()
-        workflow.outFile = getOutFile()
-
-        def result = new LinkedHashMap(5)
-        result.workflow = workflow
-        result.metrics = getMetricsList()
-        result.progress = getWorkflowProgress(false)
-        result.instant = Instant.now().toEpochMilli()
-        return result
-    }
-
-    protected Map makeHeartbeatReq() {
-        def result = new HashMap(1)
-        result.progress = getWorkflowProgress(true)
-        result.instant = Instant.now().toEpochMilli()
-        return result
-    }
-
-    protected String mapToString(def obj) {
-        if( obj == null )
-            return null
-        if( obj instanceof CharSequence )
-            return obj.toString()
-        if( obj instanceof Map ) {
-            // turn this off for multiple containers because the string representation is broken
-            return null
-        }
-        throw new IllegalArgumentException("Illegal container attribute type: ${obj.getClass().getName()} = ${obj}" )
-    }
-
-    protected Map makeTaskMap0(TraceRecord trace) {
-        Map<String,?> record = new LinkedHashMap<>(trace.store.size())
-        for( Map.Entry<String,Object> entry : trace.store.entrySet() ) {
-            def name = entry.key
-            // remove '%' char from field prefix
-            if( name.startsWith('%') )
-                name = 'p' + name.substring(1)
-            // normalise to camelCase
-            name = underscoreToCamelCase(name)
-            // put the value
-            record.put(name, fixTaskField(name,entry.value))
-        }
-
-        // prevent invalid tag data
-        if( record.tag!=null && !(record.tag instanceof CharSequence)) {
-            final msg = "Invalid tag value for process: ${record.process} -- A string is expected instead of type: ${record.tag.getClass().getName()}; offending value=${record.tag}"
-            log.warn1(msg, cacheKey: record.process)
-            record.tag = null
-        }
-
-        // add transient fields
-        record.executor = trace.getExecutorName()
-        record.cloudZone = trace.getMachineInfo()?.zone
-        record.machineType = trace.getMachineInfo()?.type
-        record.priceModel = trace.getMachineInfo()?.priceModel?.toString()
-        record.numSpotInterruptions = trace.getNumSpotInterruptions()
-        record.logStreamId = trace.getLogStreamId()
-
-        return record
-    }
-
-
-    static protected Object fixTaskField(String name, value) {
-        if( TraceRecord.FIELDS[name] == 'date' )
-            return value ? OffsetDateTime.ofInstant(Instant.ofEpochMilli(value as long), ZoneId.systemDefault()) : null
-        else
-            return value
-    }
-
-    protected Map makeTasksReq(Collection<TraceRecord> tasks) {
-
-        def payload = new ArrayList(tasks.size())
-        for( TraceRecord rec : tasks ) {
-            payload << makeTaskMap0(rec)
-        }
-
-        final result = new LinkedHashMap(5)
-        result.put('tasks', payload)
-        result.put('progress', getWorkflowProgress(true))
-        result.put('containers', getNewContainers(tasks))
-        result.instant = Instant.now().toEpochMilli()
-        return result
-    }
-
-    protected List<ContainerMeta> getNewContainers(Collection<TraceRecord> tasks) {
-        final result = new ArrayList<ContainerMeta>()
-        for( TraceRecord it : tasks ) {
-            final meta = it.getContainerMeta()
-            if( meta && !allContainers.get(meta.targetImage) ) {
-                allContainers.put(meta.targetImage, Boolean.TRUE)
-                result.add(meta)
-            }
-        }
-        return result
-    }
-
-    protected List getMetricsList() {
-        return aggregator.computeSummaryList()
-    }
-
-    protected WorkflowProgress getWorkflowProgress(boolean quick) {
-        def stats = quick ? session.getStatsObserver().getQuickStats() : session.getStatsObserver().getStats()
-        new WorkflowProgress(stats)
-    }
-
     /**
      * Little helper function that can be called for logging upon an incoming HTTP response
      */
@@ -774,18 +377,6 @@ class TowerClient implements TraceObserverV2 {
         }
     }
 
-    protected String underscoreToCamelCase(String str) {
-        if( !str.contains('_') )
-            return str
-
-        final words = str.tokenize('_')
-        def result = words[0]
-        for( int i=1; i<words.size(); i++ )
-            result+=words[i].capitalize()
-
-        return result
-    }
-
 
     protected Map<String,Integer> loadSchema() {
         final props = new Properties()
@@ -796,51 +387,6 @@ class TowerClient implements TraceObserverV2 {
             result.put( key, value ? value as Integer : null )
         }
         return result
-    }
-
-    protected void sendTasks0(dummy) {
-        final tasks = new HashMap<TaskId, TraceRecord>(TASKS_PER_REQUEST)
-        boolean complete = false
-        long previous = System.currentTimeMillis()
-        final long period = requestInterval.millis
-        final long delay = period / 10 as long
-
-        while( !complete ) {
-            final ProcessEvent ev = events.poll(delay, TimeUnit.MILLISECONDS)
-            // reconcile task events ie. send out only the last event
-            if( ev ) {
-                log.trace "Tower event=$ev"
-                if( ev.trace )
-                    tasks[ev.trace.taskId] = ev.trace
-                if( ev.completed )
-                    complete = true
-            }
-
-            // check if there's something to send
-            final now = System.currentTimeMillis()
-            final delta = now -previous
-
-            if( !tasks ) {
-                if( delta > aliveInterval.millis ) {
-                    final req = makeHeartbeatReq()
-                    final resp = sendHttpMessage(urlTraceHeartbeat, req, 'PUT')
-                    logHttpResponse(urlTraceHeartbeat, resp)
-                    previous = now
-                }
-                continue
-            }
-
-            if( delta > period || tasks.size() >= TASKS_PER_REQUEST || complete ) {
-                // send
-                final req = makeTasksReq(tasks.values())
-                final resp = sendHttpMessage(urlTraceProgress, req, 'PUT')
-                logHttpResponse(urlTraceProgress, resp)
-
-                // clean up for next iteration
-                previous = now
-                tasks.clear()
-            }
-        }
     }
 
 }
