--- plugins/nf-tower/src/main/io/seqera/tower/plugin/dataset/WorkspaceOrgDto.groovy
@@ -1,35 +1,0 @@
-/*
- * Copyright 2013-2026, Seqera Labs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.seqera.tower.plugin.dataset
-
-import groovy.transform.CompileStatic
-import groovy.transform.ToString
-
-/**
- * DTO representing an org+workspace entry returned by GET /user/{userId}/workspaces
- *
- * @author Seqera Labs
- */
-@CompileStatic
-@ToString(includeNames = true)
-class WorkspaceOrgDto {
-    long orgId
-    String orgName
-    long workspaceId
-    String workspaceName
-    String workspaceFullName
-}
