+++ plugins/nf-tower/src/test/io/seqera/tower/plugin/TowerClientTest.groovy
@@ -17,25 +17,10 @@
 package io.seqera.tower.plugin
 
 import java.net.http.HttpResponse
-import java.nio.file.Files
 import java.time.Instant
-import java.time.OffsetDateTime
-import java.time.ZoneId
 
 import io.seqera.http.HxClient
-import nextflow.Session
-import nextflow.cloud.types.CloudMachineInfo
-import nextflow.cloud.types.PriceModel
-import nextflow.container.DockerConfig
-import nextflow.container.resolver.ContainerMeta
 import nextflow.exception.AbortOperationException
-import nextflow.script.PlatformMetadata
-import nextflow.script.ScriptBinding
-import nextflow.script.WorkflowMetadata
-import nextflow.trace.TraceRecord
-import nextflow.trace.WorkflowStats
-import nextflow.trace.WorkflowStatsObserver
-import nextflow.util.ProcessHelper
 import spock.lang.Specification
 /**
  *
@@ -65,57 +50,6 @@ class TowerClientTest extends Specification {
         then:
         thrown(Exception)
     }
-
-    def 'should create message map' () {
-        given:
-        def session = Mock(Session)
-        def params = new ScriptBinding.ParamsMap(x: "hello")
-        def meta = Mock(WorkflowMetadata)
-
-        def tower = Spy(TowerClient)
-        tower.@runName = session.runName
-        tower.@workflowId = '12ef'
-
-        when:
-        def map = tower.makeCompleteReq(session)
-        then:
-        1 * session.getWorkflowMetadata() >> meta
-        1 * session.getParams() >> params
-        1 * meta.toMap() >> [foo:1, bar:2, container: [p1: 'c1', p2: 'c2']]
-        1 * tower.getMetricsList() >> [[process:'foo', cpu: [min: 1, max:5], time: [min: 6, max: 9]]]
-        1 * tower.getWorkflowProgress(false) >> new WorkflowProgress()
-        1 * tower.getOutFile() >> 'bar.out'
-        1 * tower.getLogFile() >> 'foo.out'
-        1 * tower.getOperationId() >> 'op-12345'
-        then:
-        map.workflow.foo == 1
-        map.workflow.bar == 2
-        map.workflow.id == '12ef'
-        map.workflow.params == [x: 'hello']
-        map.workflow.container == null
-        map.metrics == [[process:'foo', cpu: [min: 1, max:5], time: [min: 6, max: 9]]]
-        map.progress == new WorkflowProgress()
-        and:
-        aroundNow(map.instant)
-        and:
-        map.workflow.outFile == 'bar.out'
-        map.workflow.logFile == 'foo.out'
-        map.workflow.operationId == 'op-12345'
-    }
-
-    def 'should capitalise underscores' () {
-        given:
-        def tower = new TowerClient()
-
-        expect:
-        tower.underscoreToCamelCase(STR) == EXPECTED
-        where:
-        STR         | EXPECTED
-        'abc'       | 'abc'
-        'a_b_c'     | 'aBC'
-        'foo__bar'  | 'fooBar'
-    }
-
 
     def 'should validate URL' () {
         given:
@@ -152,322 +86,37 @@ class TowerClientTest extends Specification {
     }
 
     def 'should get access token' () {
-        given:
-        def session = Mock(Session)
-
         when:
         def config = new TowerConfig([accessToken: 'abc'], [TOWER_ACCESS_TOKEN: 'xyz'])
-        def observer = new TowerClient(session, config)
+        def client = new TowerClient(config, [TOWER_ACCESS_TOKEN: 'xyz'])
         then:
         // the token in the config overrides the one in the env
-        observer.getAccessToken() == 'abc'
+        client.getAccessToken() == 'abc'
 
         when:
         config = new TowerConfig([accessToken: 'abc'], [TOWER_ACCESS_TOKEN: 'xyz', TOWER_WORKFLOW_ID: '111222333'])
-        observer = new TowerClient(session, config)
+        client = new TowerClient(config, [TOWER_ACCESS_TOKEN: 'xyz', TOWER_WORKFLOW_ID: '111222333'])
         then:
         // the token from the env is taken because is a tower launch aka TOWER_WORKFLOW_ID is set
-        observer.getAccessToken() == 'xyz'
+        client.getAccessToken() == 'xyz'
 
         when:
         config = new TowerConfig([:], [TOWER_ACCESS_TOKEN: 'xyz'])
-        observer = new TowerClient(session, config)
+        client = new TowerClient(config, [TOWER_ACCESS_TOKEN: 'xyz'])
         then:
-        observer.getAccessToken() == 'xyz'
+        client.getAccessToken() == 'xyz'
 
         when:
-        config = new TowerConfig([:], [:])
-        observer = new TowerClient(session, config)
-        observer.getAccessToken()
+        def c = new TowerClient()
+        c.getAccessToken()
         then:
         thrown(AbortOperationException)
-    }
-
-    def 'should post task records' () {
-        given:
-        def URL = 'http://foo.com'
-        def PROGRESS = Mock(WorkflowProgress) { getRunning()>>1; getSucceeded()>>2; getFailed()>>3 }
-        def client = Mock(HxClient)
-        def observer = Spy(TowerClient)
-        observer.@httpClient = client
-        observer.@workflowId = 'xyz-123'
-
-        def nowTs = System.currentTimeMillis()
-        def submitTs = nowTs-2000
-        def startTs = nowTs-1000
-
-        def trace = new TraceRecord([
-                taskId: 10,
-                process: 'foo',
-                workdir: "/work/dir",
-                cpus: 1,
-                submit: submitTs,
-                start: startTs,
-                complete: nowTs ])
-        trace.executorName= 'batch'
-        trace.machineInfo = new CloudMachineInfo('m4.large', 'eu-west-1b', PriceModel.spot)
-        trace.containerMeta = new ContainerMeta(requestId: '12345', sourceImage: 'ubuntu:latest', targetImage: 'wave.io/12345/ubuntu:latest')
-        when:
-        def req = observer.makeTasksReq([trace])
-        then:
-        observer.getWorkflowProgress(true) >> PROGRESS
-        and:
-        req.tasks[0].taskId == 10
-        req.tasks[0].process == 'foo'
-        req.tasks[0].workdir == "/work/dir"
-        req.tasks[0].cpus == 1
-        req.tasks[0].submit == OffsetDateTime.ofInstant(Instant.ofEpochMilli(submitTs), ZoneId.systemDefault())
-        req.tasks[0].start == OffsetDateTime.ofInstant(Instant.ofEpochMilli(startTs), ZoneId.systemDefault())
-        req.tasks[0].executor == 'batch'
-        req.tasks[0].machineType == 'm4.large'
-        req.tasks[0].cloudZone == 'eu-west-1b'
-        req.tasks[0].priceModel == 'spot'
-        and:
-        req.progress.running == 1
-        req.progress.succeeded == 2
-        req.progress.failed == 3
-        and:
-        req.containers[0].requestId == '12345'
-        req.containers[0].sourceImage == 'ubuntu:latest'
-        req.containers[0].targetImage == 'wave.io/12345/ubuntu:latest'
-        and:
-        aroundNow(req.instant)
-
-        when:
-        observer.sendHttpMessage(URL, req)
-        then:
-        1 * client.sendAsString(_) >> Mock(HttpResponse)
-
-    }
-
-
-    static now_millis = System.currentTimeMillis()
-    static now_instant = OffsetDateTime.ofInstant(Instant.ofEpochMilli(now_millis), ZoneId.systemDefault())
-
-    def 'should fix field types' () {
-
-        expect:
-        TowerClient.fixTaskField(FIELD,VALUE) == EXPECTED
-
-        where:
-        FIELD       | VALUE         | EXPECTED
-        'foo'       | 'hola'        | 'hola'
-        'submit'    | now_millis    | now_instant
-        'start'     | now_millis    | now_instant
-        'complete'  | now_millis    | now_instant
-        'complete'  | 0             | null
-    }
-
-
-    def 'should create workflow json' () {
-
-        given:
-        def sessionId = UUID.randomUUID()
-        def dir = Files.createTempDirectory('test')
-        def http = Mock(HxClient)
-        TowerClient client = Spy(new TowerClient([httpClient: http, env: ENV]))
-        and:
-        client.getOperationId() >> 'op-112233'
-        client.getLogFile() >> 'log.file'
-        client.getOutFile() >> 'out.file'
-
-        and:
-        def session = Mock(Session)
-        session.getUniqueId() >> sessionId
-        session.getRunName() >> 'foo'
-        session.config >> [:]
-        session.containerConfig >> new DockerConfig([:])
-        session.getParams() >> new ScriptBinding.ParamsMap([foo:'Hello', bar:'World'])
-
-        def meta = new WorkflowMetadata(
-                session: session,
-                projectName: 'the-project-name',
-                repository: 'git://repo.com/foo' )
-        session.getWorkflowMetadata() >> meta
-        session.getStatsObserver() >> Mock(WorkflowStatsObserver) { getStats() >> new WorkflowStats() }
-
-        when:
-        def req1 = client.makeCreateReq(session)
-        then:
-        req1.sessionId == sessionId.toString()
-        req1.runName == 'foo'
-        req1.projectName == 'the-project-name'
-        req1.repository == 'git://repo.com/foo'
-        req1.workflowId == WORKFLOW_ID
-        and:
-        aroundNow(req1.instant)
-
-        when:
-        def req = client.makeBeginReq(session)
-        then:
-        client.getWorkflowId() >> '12345'
-        and:
-        req.workflow.id == '12345'
-        req.workflow.params == [foo:'Hello', bar:'World']
-        req.workflow.outFile == 'out.file'
-        req.workflow.logFile == 'log.file'
-        req.workflow.operationId == 'op-112233'
-        and:
-        req.towerLaunch == TOWER_LAUNCH
-        and:
-        aroundNow(req.instant)
-
-        cleanup:
-        dir?.deleteDir()
-
-        where:
-        ENV                         | WORKFLOW_ID   | TOWER_LAUNCH
-        [:]                         | null          | false
-        [TOWER_WORKFLOW_ID: '1234'] | '1234'        | true
-
-    }
-
-    def 'should convert map' () {
-        given:
-        def tower = new TowerClient()
-
-        expect:
-        tower.mapToString(null)  == null
-        tower.mapToString('ciao') == 'ciao'
-        tower.mapToString([:]) == null
-        tower.mapToString([p:'foo', q:'bar']) == null
-    }
-
-
-    def 'should load schema col len' () {
-        given:
-        def tower = new TowerClient()
-
-        when:
-        def schema = tower.loadSchema()
-        then:
-        schema.get('workflow.start')  == null
-        schema.get('workflow.profile') == 100
-        schema.get('workflow.projectDir') == 255
-    }
-
-    def 'should create init request' () {
-        given:
-        def uuid = UUID.randomUUID()
-        def client = new TowerClient(env: [TOWER_WORKFLOW_ID: 'x123'])
-        def meta = Mock(WorkflowMetadata) {
-            getProjectName() >> 'the-project-name'
-            getRepository() >> 'git://repo.com/foo'
-        }
-        def session = Mock(Session) {
-            getUniqueId() >> uuid
-            getRunName() >> 'foo_bar'
-            getWorkflowMetadata() >> meta
-        }
-
-        when:
-        def req = client.makeCreateReq(session)
-        then:
-        req.sessionId == uuid.toString()
-        req.runName == 'foo_bar'
-        req.projectName == 'the-project-name'
-        req.repository == 'git://repo.com/foo'
-        req.workflowId == 'x123'
-        and:
-        aroundNow(req.instant)
-
-        and:
-        client.towerLaunch
-    }
-
-    def 'should post create request' () {
-        given:
-        def uuid = UUID.randomUUID()
-        def platform = new PlatformMetadata()
-        def meta = Mock(WorkflowMetadata) {
-            getProjectName() >> 'the-project-name'
-            getRepository() >> 'git://repo.com/foo'
-            getPlatform() >> platform
-        }
-        def session = Mock(Session) {
-            getUniqueId() >> uuid
-            getRunName() >> 'foo_bar'
-            getWorkflowMetadata() >> meta
-        }
-        def config = new TowerConfig([:], [:])
-
-        def client = Spy(new TowerClient(session, config))
-
-        when:
-        client.onFlowCreate(session)
-        then:
-        1 * client.makeCreateReq(session) >> [runName: 'foo']
-        1 * client.sendHttpMessage('https://api.cloud.seqera.io/trace/create', [runName: 'foo'], 'POST') >> new TowerClient.Response(200, '{"workflowId":"xyz123","watchUrl":"https://cloud.seqera.io/watch/xyz123"}')
-        and:
-        client.runName == 'foo_bar'
-        client.runId == uuid.toString()
-        and:
-        client.workflowId == 'xyz123'
-        client.@watchUrl == 'https://cloud.seqera.io/watch/xyz123'
-        !client.towerLaunch
-        and:
-        platform.workflowId == 'xyz123'
-        platform.workflowUrl == 'https://cloud.seqera.io/watch/xyz123'
-
-    }
-
-    def 'should set workflowUrl on platform metadata during onFlowBegin' () {
-        given:
-        def platform = new PlatformMetadata()
-        def meta = Mock(WorkflowMetadata) {
-            getPlatform() >> platform
-        }
-        def session = Mock(Session) {
-            getWorkflowMetadata() >> meta
-        }
-        def config = new TowerConfig([:], [:])
-        def client = Spy(new TowerClient(session, config))
-        client.@workflowId = 'abc123'
-
-        when:
-        client.onFlowBegin()
-        then:
-        1 * client.makeBeginReq(session) >> [foo: 'bar']
-        1 * client.sendHttpMessage(_, [foo: 'bar'], 'PUT') >> new TowerClient.Response(200, '{"watchUrl":"https://cloud.seqera.io/watch/abc123"}')
-        and:
-        client.@watchUrl == 'https://cloud.seqera.io/watch/abc123'
-        platform.workflowUrl == 'https://cloud.seqera.io/watch/abc123'
-    }
-
-    def 'should get trace endpoint' () {
-        given:
-        def config = new TowerConfig([:], [:])
-        def tower = new TowerClient(Mock(Session), config)
-        tower.workflowId = '12345'
-
-        expect:
-        tower.getUrlTraceCreate() == 'https://api.cloud.seqera.io/trace/create'
-        tower.getUrlTraceBegin() == 'https://api.cloud.seqera.io/trace/12345/begin'
-        tower.getUrlTraceProgress() == 'https://api.cloud.seqera.io/trace/12345/progress'
-        tower.getUrlTraceHeartbeat() == 'https://api.cloud.seqera.io/trace/12345/heartbeat'
-        tower.getUrlTraceComplete() == 'https://api.cloud.seqera.io/trace/12345/complete'
-    }
-
-    def 'should get trace endpoint with workspace' () {
-        given:
-        def config = new TowerConfig([workspaceId: '300'], [:])
-        def tower = new TowerClient(Mock(Session), config)
-        tower.workflowId = '12345'
-
-        expect:
-        tower.getUrlTraceCreate() == 'https://api.cloud.seqera.io/trace/create?workspaceId=300'
-        tower.getUrlTraceBegin() == 'https://api.cloud.seqera.io/trace/12345/begin?workspaceId=300'
-        tower.getUrlTraceProgress() == 'https://api.cloud.seqera.io/trace/12345/progress?workspaceId=300'
-        tower.getUrlTraceHeartbeat() == 'https://api.cloud.seqera.io/trace/12345/heartbeat?workspaceId=300'
-        tower.getUrlTraceComplete() == 'https://api.cloud.seqera.io/trace/12345/complete?workspaceId=300'
     }
 
     def 'should set the auth token' () {
         given:
         def http = Mock(HxClient.Builder)
-        def session = Mock(Session)
-        def config = new TowerConfig([:], [:])
-        def client = Spy(new TowerClient(session, config))
+        def client = new TowerClient()
         and:
         def SIMPLE = '4ffbf1009ebabea77db3d72efefa836dfbb71271'
         def BEARER = 'eyJ0aWQiOiA1fS5jZmM1YjVhOThjZjM2MTk1NjBjZWU1YmMwODUxYzA1ZjkzMDdmN2Iz'
@@ -490,170 +139,69 @@ class TowerClientTest extends Specification {
         1 * http.refreshTokenUrl(_) >> http
     }
 
-    def 'should fetch workflow meta' () {
+    def 'should get trace endpoint' () {
         given:
-        def client = Spy(new TowerClient(env: ENV))
+        def client = new TowerClient()
+        client.@endpoint = TowerClient.DEF_ENDPOINT_URL
 
         expect:
-        client.getOperationId() == OP_ID
-        client.getLogFile() == LOG_FILE
-        client.getOutFile() == OUT_FILE
-
-        where:
-        OP_ID                                           | OUT_FILE      | LOG_FILE    | ENV
-        null                                            | null          | null        | [:]
-        "local-platform::${ProcessHelper.selfPid()}"    | null          | null        | [TOWER_ALLOW_NEXTFLOW_LOGS:'true']
-        'aws-batch::1234z'                              | 'xyz.out'     | 'hola.log'  | [TOWER_ALLOW_NEXTFLOW_LOGS:'true', AWS_BATCH_JOB_ID: '1234z', NXF_OUT_FILE: 'xyz.out', NXF_LOG_FILE: 'hola.log']
+        client.getUrlTraceCreate(null) == 'https://api.cloud.seqera.io/trace/create'
+        client.getUrlTraceBegin(null, '12345') == 'https://api.cloud.seqera.io/trace/12345/begin'
+        client.getUrlTraceProgress(null, '12345') == 'https://api.cloud.seqera.io/trace/12345/progress'
+        client.getUrlTraceHeartbeat(null, '12345') == 'https://api.cloud.seqera.io/trace/12345/heartbeat'
+        client.getUrlTraceComplete(null, '12345') == 'https://api.cloud.seqera.io/trace/12345/complete'
     }
 
-    def 'should deduplicate containers' () {
+    def 'should get trace endpoint with workspace' () {
         given:
-        def client = Spy(new TowerClient())
-        and:
-        def c1 = new ContainerMeta(requestId: '12345', sourceImage: 'ubuntu:latest', targetImage: 'wave.io/12345/ubuntu:latest')
-        def c2 = new ContainerMeta(requestId: '54321', sourceImage: 'ubuntu:latest', targetImage: 'wave.io/54321/ubuntu:latest')
-        and:
-        def trace1 = new TraceRecord(
-                taskId: 1,
-                process: 'foo',
-                workdir: "/work/dir",
-                cpus: 1,
-                submit: System.currentTimeMillis(),
-                start: System.currentTimeMillis(),
-                complete: System.currentTimeMillis())
-        trace1.containerMeta = c1
-        and:
-        def trace2 = new TraceRecord(
-            taskId: 2,
-            process: 'foo',
-            workdir: "/work/dir",
-            cpus: 1,
-            submit: System.currentTimeMillis(),
-            start: System.currentTimeMillis(),
-            complete: System.currentTimeMillis())
-        trace2.containerMeta = c2
-        and:
-        def trace3 = new TraceRecord(
-            taskId: 3,
-            process: 'foo',
-            workdir: "/work/dir",
-            cpus: 1,
-            submit: System.currentTimeMillis(),
-            start: System.currentTimeMillis(),
-            complete: System.currentTimeMillis())
-        trace3.containerMeta = c2
+        def client = new TowerClient()
+        client.@endpoint = TowerClient.DEF_ENDPOINT_URL
 
         expect:
-        client.getNewContainers([trace1]) == [c1]
-        and:
-        client.getNewContainers([trace1]) == []
-        and:
-        client.getNewContainers([trace1, trace2, trace3]) == [c2]
+        client.getUrlTraceCreate('300') == 'https://api.cloud.seqera.io/trace/create?workspaceId=300'
+        client.getUrlTraceBegin('300', '12345') == 'https://api.cloud.seqera.io/trace/12345/begin?workspaceId=300'
+        client.getUrlTraceProgress('300', '12345') == 'https://api.cloud.seqera.io/trace/12345/progress?workspaceId=300'
+        client.getUrlTraceHeartbeat('300', '12345') == 'https://api.cloud.seqera.io/trace/12345/heartbeat?workspaceId=300'
+        client.getUrlTraceComplete('300', '12345') == 'https://api.cloud.seqera.io/trace/12345/complete?workspaceId=300'
     }
 
-    def 'should not send complete request when onFlowBegin was not invoked' () {
+    def 'should load schema col len' () {
         given:
-        def client = Spy(new TowerClient())
-        client.@workflowId = 'xyz-123'
-        client.@sender = null
-        client.@reports = Mock(TowerReports)
+        def tower = new TowerClient()
 
         when:
-        client.onFlowComplete()
-
+        def schema = tower.loadSchema()
         then:
-        1 * client.@reports.publishRuntimeReports()
-        1 * client.@reports.flowComplete()
-        0 * client.sendHttpMessage(_, _, _)
+        schema.get('workflow.start')  == null
+        schema.get('workflow.profile') == 100
+        schema.get('workflow.projectDir') == 255
     }
 
     def 'should handle HTTP request with content'() {
-        given: 'a TowerClient'
+        given:
         def tower = new TowerClient()
         def content = '{"test": "data"}'
         def request = tower.makeRequest('http://example.com/test', content, 'POST')
 
-        expect: 'the request should be created with the content'
+        expect:
         request != null
         request.method() == 'POST'
         request.uri().toString() == 'http://example.com/test'
     }
 
-    def 'should include numSpotInterruptions in task map'() {
+    def 'should send http message' () {
         given:
-        def client = Spy(new TowerClient())
-        client.getWorkflowProgress(true) >> new WorkflowProgress()
-
-        def now = System.currentTimeMillis()
-        def trace = new TraceRecord([
-            taskId: 42,
-            process: 'foo',
-            workdir: "/work/dir",
-            cpus: 1,
-            submit: now-2000,
-            start: now-1000,
-            complete: now
-        ])
-        trace.setNumSpotInterruptions(3)
+        def client = Mock(HxClient)
+        def tower = new TowerClient()
+        tower.@httpClient = client
 
         when:
-        def req = client.makeTasksReq([trace])
-
+        def resp = tower.sendHttpMessage('http://foo.com', [foo: 'bar'], 'POST')
         then:
-        req.tasks.size() == 1
-        req.tasks[0].numSpotInterruptions == 3
-    }
-
-    def 'should include logStreamId in task map'() {
-        given:
-        def client = Spy(new TowerClient())
-        client.getWorkflowProgress(true) >> new WorkflowProgress()
-
-        def now = System.currentTimeMillis()
-        def trace = new TraceRecord([
-            taskId: 42,
-            process: 'foo',
-            workdir: "/work/dir",
-            cpus: 1,
-            submit: now-2000,
-            start: now-1000,
-            complete: now
-        ])
-        trace.setLogStreamId('arn:aws:logs:us-east-1:123456789:log-group:/ecs/task:log-stream:abc123')
-
-        when:
-        def req = client.makeTasksReq([trace])
-
-        then:
-        req.tasks.size() == 1
-        req.tasks[0].logStreamId == 'arn:aws:logs:us-east-1:123456789:log-group:/ecs/task:log-stream:abc123'
-    }
-
-    def 'should include accelerator request in task map'() {
-        given:
-        def client = Spy(new TowerClient())
-        client.getWorkflowProgress(true) >> new WorkflowProgress()
-
-        def now = System.currentTimeMillis()
-        def trace = new TraceRecord([
-            taskId: 42,
-            process: 'foo',
-            workdir: "/work/dir",
-            cpus: 1,
-            submit: now-2000,
-            start: now-1000,
-            complete: now,
-            accelerator: 2,
-            acceleratorType: 'v100'
-        ])
-
-        when:
-        def req = client.makeTasksReq([trace])
-
-        then:
-        req.tasks.size() == 1
-        req.tasks[0].accelerator == 2
-        req.tasks[0].acceleratorType == 'v100'
+        1 * client.sendAsString(_) >> Mock(HttpResponse) { statusCode() >> 200; body() >> '{}' }
+        and:
+        !resp.error
+        resp.code == 200
     }
 
 }
