+++ plugins/nf-tower/src/test/io/seqera/tower/plugin/TowerFusionEnvTest.groovy
@@ -16,6 +16,8 @@
 
 package io.seqera.tower.plugin
 
+import nextflow.exception.AbortOperationException
+
 import static com.github.tomakehurst.wiremock.client.WireMock.*
 
 import java.time.Instant
@@ -70,6 +72,7 @@ class TowerFusionEnvTest extends Specification {
 
     def 'should return the endpoint from the config'() {
         given: 'a session'
+        SysEnv.push(['TOWER_API_ENDPOINT': 'https://tower.nf', 'TOWER_ACCESS_TOKEN': 'abc123'])
         Global.session = Mock(Session) {
             config >> [
                 tower: [
@@ -83,11 +86,13 @@ class TowerFusionEnvTest extends Specification {
 
         then: 'the endpoint has the expected value'
         provider.endpoint == 'https://tower.nf'
+        cleanup:
+        SysEnv.pop()
     }
 
     def 'should return the endpoint from the environment'() {
         setup:
-        SysEnv.push(['TOWER_API_ENDPOINT': 'https://tower.nf'])
+        SysEnv.push(['TOWER_API_ENDPOINT': 'https://tower.nf', 'TOWER_ACCESS_TOKEN': 'abc123'])
         Global.session = Mock(Session) {
             config >> [:]
         }
@@ -103,6 +108,7 @@ class TowerFusionEnvTest extends Specification {
     }
 
     def 'should return the default endpoint'() {
+        SysEnv.push(['TOWER_ACCESS_TOKEN': 'abc123'])
         when: 'session config is empty'
         Global.session = Mock(Session) {
             config >> [
@@ -177,6 +183,9 @@ class TowerFusionEnvTest extends Specification {
 
         then: 'the endpoint has the expected value'
         provider.endpoint == TowerClient.DEF_ENDPOINT_URL
+
+        cleanup:
+        SysEnv.pop()
     }
 
     def 'should return the access token from the config'() {
@@ -249,7 +258,8 @@ class TowerFusionEnvTest extends Specification {
         def provider = new TowerFusionToken()
 
         then: 'the access token has the expected value'
-        provider.accessToken == null
+        def e = thrown(AbortOperationException)
+        e.message.contains("Missing Seqera Platform access token")
 
         cleanup:
         SysEnv.pop()
@@ -307,8 +317,8 @@ class TowerFusionEnvTest extends Specification {
                 )
         )
         and:
-        def client = TowerFactory.client(session, SysEnv.get())
-        client.onFlowCreate(session)
+        def observer = new TowerFactory().create(session)[0]
+        observer.onFlowCreate(session)
 
         and: 'a mock endpoint returning a valid token'
         final now = Instant.now()
@@ -371,7 +381,7 @@ class TowerFusionEnvTest extends Specification {
                 )
         )
         and:
-        def client = TowerFactory.client(session, SysEnv.get())
+        def client = new TowerFactory().create(session)[0]
         client.onFlowCreate(session)
 
         and: 'a mock endpoint returning a valid token'
@@ -439,7 +449,7 @@ class TowerFusionEnvTest extends Specification {
                 )
         )
         and:
-        def client = TowerFactory.client(session, SysEnv.get())
+        def client = new TowerFactory().create(session)[0]
         client.onFlowCreate(session)
 
         and: 'prepare stubs'
@@ -531,7 +541,7 @@ class TowerFusionEnvTest extends Specification {
                 )
         )
         and:
-        def client = TowerFactory.client(session, SysEnv.get())
+        def client = new TowerFactory().create(session)[0]
         client.onFlowCreate(session)
         and: 'a mock endpoint returning an error'
         wireMockServer.stubFor(
